Implement sleep grace period
Use sleep as a start for a grace period:

- Kill tunnel after grace expiration.
- Cancel tunnel disposal on wake-up and reset counter.
- Default grace to 1 minute.

Keeping tunnel alive highlighted an insane amount of sleep/wake
cycles when connected via mobile data, even after disconnecting
the socket. By killing the process, resort to on-demand rules
instead to bring the tunnel back to life.

This will avoid a ridiculous amount of reconnections/negotiations
for weird sessions of just a few seconds. For some reason the
system may wake as soon as 1s after sleep, to then sleep again
shortly afterwards!