Match packetFlow lifetime in NETunnelInterface
There is an issue with how `packetFlow` and `NETunnelInterface` are
treated in terms of lifetime. Lifetime of `packetFlow` is lifetime
of `NEPacketTunnelProvider`, whereas `NETunnelInterface` is being
recreated for each new `SessionProxy` (VPN) session.

The illusion of a per-session tunnel breaks when it comes to
`NEPackeTunnelFlow.readPackets()`, because we may end up in a
situation where two VPN sessions invoke it at the same time during
a reconnection.

Given that `packetFlow`, due to its lifetime, is basically a singleton
when seen from `NETunnelInterface`, the read callback will only be
invoked in one of the two VPN sessions. If the callback returns to the
disconnecting session instead of the reconnecting one, the deferred loop
in `NETunnelInterface.loopReadPackets()` will break and die, because
`NETunnelInterface` will have been released (`weak self`).

The effect is that the link interface is up and running, but data
flowing on TUN is not read at all. High-level, this results in a network
stall.

Solution:

- Set `NETunnelInterface` wrapper once at `SessionProxy` creation.
- Retain `packetFlow` weakly, it's non-optional and survives sessions.

Fixes #2